requirements.txt
langchain==0.1.20
langchain-groq==0.1.3
langgraph==0.0.50
streamlit==1.32.0
faiss-cpu==1.8.0
sentence-transformers==2.5.1
python-dotenv==1.0.1

